fix/Incorrect_marker
According to the (inline) doc the marker is of type Hash256. Reality the marker (of at least) used in the account_lines is of type 'Hash256,int'

Solution: Introduction of a new type, here called ExtendedHash, which checks dynamically for marker (Hash256) and for the returned format 'Hash256,int'
Test run with a extended set (use the solo coin for example; 1000s of account lines) with the lines.go program shows the expected result (Previous result was a json Unmarshal error on the , in the marker).